Refactor to have relative paths in mapping to remote configs
The new gathering conditions map will reference to those files as relative paths.
We can already implement that now in preparation for the future version of the gathering conditions
as it is just a matter of storing mapping and remote configs together.

The changes include:

- Remove cluster mapping related config options.
These options are actually not leveraged and have been proved not effective to cover
other cases, e.g. being able to handle different names for different versions (as needed for the new conditions).
In addition, being the paths relative, it make sense that they are bundled together.

- Rename "RemoteConfigurationPath" to "RemoteConfigurationPath" following the same pattern as "RulesPath" in the StorageConfig struct

- Rename "remote_configuration" to "remote_configurations" in the config

- Replace fmt.Sprintf for path construction with filepath.Join

- Embedded the path construction within ClusterMapping.
The ClusterMapping only need to know the root directory (without care about whether it is canary or stable or anything else)
and use that to construct the path with the relative path.
Having the path constructed in multiple places (Storage, ClusterMapping IsValid...) is prone to errors.
There should be, ideally, only 1 place where it happens.
In addition, version is confusing in that struct as to deal with OPC versions as well.