Update building for new conditions (#370)
The conditions repo has updated its building process
and we need adapt this repo.
The changes include
- add golang to the build step for the conditions as there is a Go tool run as part of the process.
While we can have provided its binary, as we are building 2 version at the same time,
the might be differences in the tool, so to account for that we are compiling the binary at runtime.
- the new build script is in Python, so we need Python in the build image.
Also we are fixing the Python version to 3.11 and creating a venv for it following the instructions provided by the
gathering conditions repo.
- create 2 bash functions one for the new build setup and one for the legacy one